+++ Sources/SpotifyWebAPI/Object Model/Player Objects/Device.swift
@@ -10,10 +10,10 @@ public struct Device: Hashable {
     /// The device id. May be `nil`.
     public let id: String?
     
-    /// Whether the device is currently active
+    /// Whether the device is currently active.
     public let isActive: Bool
     
-    /// Whether the device is currently in a private session
+    /// Whether the device is currently in a private session.
     public let isPrivateSession: Bool
     
     /// Whether controlling this device is restricted.
@@ -23,10 +23,10 @@ public struct Device: Hashable {
     /// The name of the device.
     public let name: String
     
-    /// The type of the device
+    /// The type of the device.
     public let type: DeviceType
     
-    /// The current volume in percent.
+    /// The current volume in percent (0 to 100).
     public let volumePercent: Int?
 
 }