+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.2] - 2020-11-29
+
+### Added
+
+* `RepeatMode.cycle()` - Cycles between the different repeat modes.
+
+### Fixed
+
+* Fixed typo in `DeviceType.castAudio.rawValue`.
+
 ## [1.1.1] - 2020-11-27
 
 ### Changed